--- Assets/Scripts/switch rooms/MainMenu.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: eee196001b9018a4f910ae1e43f0d956
